patch from paul.lind@imgtec.com fix allocator alignment
Bug 11222825

Upstream chromium has a fix for this in M31 but does not apply to M30,
so creating a spot-fix in the klp branch. #ifdef for MIPS to reduce risk
as no other CPU is known to require this fix (it should be valid on all
platforms).

Change-Id: I924bea66eb3cef725a0dac89f5717a31fb471194
(cherry picked from commit fe6262e1acba725d438cd0f8464e7faaeb886b0a)